requirements.txt
opencv-python
numpy
Pillow
rich
